videothumbs_project/deploy/requirements.txt
Django==1.6
django-extensions==1.2.5
django-quicky==0.6.4
django-redis-sessions==0.4.0
path.py==5.0
pytest==2.4.2
redis==2.8.0
fabtools==0.16.0
django-celery==3.0.23
django-simple-email-confirmation==0.10.2
django-rosetta==0.7.2
django-celery-email==1.0.4
django-crispy-forms==1.4.0
South==0.8.4
ProxyTypes==0.9
pytest-django==2.4
django-redis==3.3